+++ CHANGELOG.md
@@ -2,6 +2,7 @@ Important changes to data models, configuration, and migrations between each
 AppEngine version, listed here to ease deployment and troubleshooting.
 
 ## Next Release (replace with git tag when deployed)
+ * Note: Reverted public `cache-control` headers except for `/documentation/` pages.
 
 ## `20251023t081900-all`
  * Bump runtimeVersion to `2025.10.22`.